recommendation-service/requirements.txt
requests~=2.32.3
tensorflow==2.19.0
pandas~=2.2.3
pip-system-certs~=4.0
psycopg2-binary~=2.9.10
uvicorn~=0.32.0
fastapi~=0.115.12
starlette<=0.47.0
pydantic~=2.11.6
numpy<=2.2.0
confluent-kafka==2.10.1
schedule~=1.2.2
py-eureka-client~=0.11.13